Safari-611.1.10/LayoutTests/fast/animation/request-animation-frame-throttle-inside-overflow-scroll-expected.txt

Tests that requestAnimationFrame is throttled in subframes are inside scrolled overflow scroll

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Frame is initially inside the viewport so requestAnimationFrame should not be throttled
PASS testFrame.contentWindow.internals.requestAnimationFrameThrottlingReasons() is "[Unthrottled]"
Scrolling overflow.
requestAnimationFrame should still not be throttled
PASS testFrame.contentWindow.internals.requestAnimationFrameThrottlingReasons() is "[Unthrottled]"
PASS successfullyParsed is true

TEST COMPLETE
